M1592 refactor:update language and clear size modal.
